Remove serialization code
Binary serialization is not supported in CoreCLR and hence must be removed
from the subset which runs there.  This code is only used to support
CodeLens so rather than moving it around we are simply removing it
completely.  It will be relocated to CodeLens internally to continue supporting
their scenarios.

The Public API changes to DocumentId and ProjectId were necessary to
continue supporting serialization outside of the Roslyn code base.

closes #915
