+++ src/Compilers/CSharp/Desktop/PublicAPI.txt
@@ -4,20 +4,10 @@ Microsoft.CodeAnalysis.CSharp.CSharpCommandLineArguments.ParseOptions.get -> Mic
 Microsoft.CodeAnalysis.CSharp.CSharpCommandLineParser
 Microsoft.CodeAnalysis.CSharp.CSharpCommandLineParser.Parse(System.Collections.Generic.IEnumerable<string> args, string baseDirectory, string additionalReferencePaths = null) -> Microsoft.CodeAnalysis.CSharp.CSharpCommandLineArguments
 Microsoft.CodeAnalysis.CSharp.CSharpFileSystemExtensions
-Microsoft.CodeAnalysis.CSharp.CSharpSerializableCompilationOptions
-Microsoft.CodeAnalysis.CSharp.CSharpSerializableCompilationOptions.CSharpSerializableCompilationOptions(Microsoft.CodeAnalysis.CSharp.CSharpCompilationOptions options) -> void
-Microsoft.CodeAnalysis.CSharp.CSharpSerializableCompilationOptions.Options.get -> Microsoft.CodeAnalysis.CSharp.CSharpCompilationOptions
-Microsoft.CodeAnalysis.CSharp.CSharpSerializableParseOptions
-Microsoft.CodeAnalysis.CSharp.CSharpSerializableParseOptions.CSharpSerializableParseOptions(Microsoft.CodeAnalysis.CSharp.CSharpParseOptions options) -> void
-Microsoft.CodeAnalysis.CSharp.CSharpSerializableParseOptions.Options.get -> Microsoft.CodeAnalysis.ParseOptions
 override Microsoft.CodeAnalysis.CSharp.CSharpCommandLineArguments.CompilationOptionsCore.get -> Microsoft.CodeAnalysis.CompilationOptions
 override Microsoft.CodeAnalysis.CSharp.CSharpCommandLineArguments.ParseOptionsCore.get -> Microsoft.CodeAnalysis.ParseOptions
 override Microsoft.CodeAnalysis.CSharp.CSharpCommandLineParser.RegularFileExtension.get -> string
 override Microsoft.CodeAnalysis.CSharp.CSharpCommandLineParser.ScriptFileExtension.get -> string
-override Microsoft.CodeAnalysis.CSharp.CSharpSerializableCompilationOptions.CommonOptions.get -> Microsoft.CodeAnalysis.CompilationOptions
-override Microsoft.CodeAnalysis.CSharp.CSharpSerializableCompilationOptions.GetObjectData(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context) -> void
-override Microsoft.CodeAnalysis.CSharp.CSharpSerializableParseOptions.CommonOptions.get -> Microsoft.CodeAnalysis.ParseOptions
-override Microsoft.CodeAnalysis.CSharp.CSharpSerializableParseOptions.GetObjectData(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context) -> void
 static Microsoft.CodeAnalysis.CSharp.CSharpCommandLineParser.ParseConditionalCompilationSymbols(string value, out System.Collections.Generic.IEnumerable<Microsoft.CodeAnalysis.Diagnostic> diagnostics) -> System.Collections.Generic.IEnumerable<string>
 static Microsoft.CodeAnalysis.CSharp.CSharpFileSystemExtensions.Emit(this Microsoft.CodeAnalysis.CSharp.CSharpCompilation compilation, string outputPath, string pdbPath = null, string xmlDocumentationPath = null, string win32ResourcesPath = null, System.Collections.Generic.IEnumerable<Microsoft.CodeAnalysis.ResourceDescription> manifestResources = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> Microsoft.CodeAnalysis.Emit.EmitResult
 static readonly Microsoft.CodeAnalysis.CSharp.CSharpCommandLineParser.Default -> Microsoft.CodeAnalysis.CSharp.CSharpCommandLineParser
