+++ src/Compilers/Core/Desktop/CodeAnalysis.Desktop.csproj
@@ -97,9 +97,6 @@
     <Compile Include="MetadataFileReferenceResolver.cs" />
     <Compile Include="PathKind.cs" />
     <Compile Include="PathUtilities.cs" />
-    <Compile Include="SerializableCompilationOptions.cs" />
-    <Compile Include="SerializableParseOptions.cs" />
-    <Compile Include="SerializationInfoExtensions.cs" />
     <Compile Include="SourceFileResolver.cs" />
     <Compile Include="SqmServiceProvider.cs" />
     <Compile Include="XmlFileResolver.cs" />
