+++ src/Compilers/VisualBasic/Desktop/PublicAPI.txt
@@ -2,20 +2,10 @@ Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineArguments
 Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineParser
 Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineParser.New(isInteractive As Boolean = False) -> Void
 Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineParser.Parse(args As System.Collections.Generic.IEnumerable(Of String), baseDirectory As String, additionalReferencePaths As String = Nothing) -> Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineArguments
-Microsoft.CodeAnalysis.VisualBasic.VisualBasicSerializableCompilationOptions
-Microsoft.CodeAnalysis.VisualBasic.VisualBasicSerializableCompilationOptions.New(options As Microsoft.CodeAnalysis.VisualBasic.VisualBasicCompilationOptions) -> Void
-Microsoft.CodeAnalysis.VisualBasic.VisualBasicSerializableCompilationOptions.Options() -> Microsoft.CodeAnalysis.VisualBasic.VisualBasicCompilationOptions
-Microsoft.CodeAnalysis.VisualBasic.VisualBasicSerializableParseOptions
-Microsoft.CodeAnalysis.VisualBasic.VisualBasicSerializableParseOptions.New(options As Microsoft.CodeAnalysis.VisualBasic.VisualBasicParseOptions) -> Void
-Microsoft.CodeAnalysis.VisualBasic.VisualBasicSerializableParseOptions.Options() -> Microsoft.CodeAnalysis.VisualBasic.VisualBasicParseOptions
 Overrides Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineArguments.CompilationOptionsCore() -> Microsoft.CodeAnalysis.CompilationOptions
 Overrides Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineArguments.ParseOptionsCore() -> Microsoft.CodeAnalysis.ParseOptions
 Overrides Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineParser.RegularFileExtension() -> String
 Overrides Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineParser.ScriptFileExtension() -> String
-Overrides Microsoft.CodeAnalysis.VisualBasic.VisualBasicSerializableCompilationOptions.CommonOptions() -> Microsoft.CodeAnalysis.CompilationOptions
-Overrides Microsoft.CodeAnalysis.VisualBasic.VisualBasicSerializableCompilationOptions.GetObjectData(info As System.Runtime.Serialization.SerializationInfo, context As System.Runtime.Serialization.StreamingContext) -> Void
-Overrides Microsoft.CodeAnalysis.VisualBasic.VisualBasicSerializableParseOptions.CommonOptions() -> Microsoft.CodeAnalysis.ParseOptions
-Overrides Microsoft.CodeAnalysis.VisualBasic.VisualBasicSerializableParseOptions.GetObjectData(info As System.Runtime.Serialization.SerializationInfo, context As System.Runtime.Serialization.StreamingContext) -> Void
 Shared Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineParser.ParseConditionalCompilationSymbols(symbolList As String, ByRef diagnostics As System.Collections.Generic.IEnumerable(Of Microsoft.CodeAnalysis.Diagnostic), symbols As System.Collections.Generic.IEnumerable(Of System.Collections.Generic.KeyValuePair(Of String, Object)) = Nothing) -> System.Collections.Generic.IReadOnlyDictionary(Of String, Object)
 Shared ReadOnly Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineParser.Default -> Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineParser
 Shared ReadOnly Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineParser.Interactive -> Microsoft.CodeAnalysis.VisualBasic.VisualBasicCommandLineParser
